move get_plugins to internal module
